LayoutTests/fast/viewport/viewport-129.html

<head>
    <title>Check that number parsing stops at the first non-number characters and still uses the numeric prefix.</title>
    <meta name="viewport" content="width=123x456">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(160, 320, 480, 320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
